+++ kit/lib/apollo.js
@@ -17,11 +17,11 @@ import { getServerURL } from 'kit/lib/env';
 // ----------------------
 
 // Helper function to create a new Apollo client, by merging in
-// passed options alongside the defaults
+// passed options alongside any set by `config.setApolloOptions` and defaults
 export function createClient(opt = {}) {
   return new ApolloClient(Object.assign({
     reduxRootSelector: state => state.apollo,
-  }, opt));
+  }, config.apolloOptions, opt));
 }
 
 // Wrap `createNetworkInterface` to attach middleware